+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [HexaPDF::Document::Layout#method_missing] for more convenient box creation
 * [HexaPDF::Document::Composer#method_missing] for more convenient box drawing
+* [HexaPDF::Document::Layout#inline_box] for easy creation of inline boxes
 * [HexaPDF::Type::OutlineItem#open?] for getting the open state of an outline
   item